Day 2 - Final Commit - Ready for Review
This project is complete and ready for review. I add cross browser extensions, styles for the background and images of the invalid/valid boxes and rules for when certain elements are focused, as well as entirely re-coded the form due to mistakes made in the first commit. I also added documentation.